Some refactoring and the introduction of a Go routine for periodic interface rate stats collection.
